@@ -2,14 +2,19 @@
 
 export const actionTypes = {
   // action types for property tree
-  updatePropertyTreeNode: 'PROPERTY_TREE_UPDATE_NODE',
-  changePropertyTreeNode: 'PROPERTY_TREE_CHANGE_NODE',
-  initializePropertyTree: 'PROPERTY_TREE_INITIALIZE',
-  startListeningToNode: 'PROPERTY_TREE_START_LISTENING',
-  stopListeningToNode: 'PROPERTY_TREE_STOP_LISTENING',
-  removeNode: 'PROPERTY_TREE_REMOVE_NODE',
-  addNode: 'PROPERTY_TREE_ADD_NODE',
-  insertNode: 'PROPERTY_TREE_INSERT_NODE',
+  updatePropertyValue: 'PROPERTY_TREE_UPDATE_NODE',
+  setPropertyValue: 'PROPERTY_TREE_SET',
+  subscribeToProperty: 'PROPERTY_TREE_SUBSCRIBE',
+  unsubscribeToProperty: 'PROPERTY_TREE_UNSUBSCRIBE',
+  addPropertyOwners: 'PROPERTY_TREE_ADD_OWNERS',
+  addProperties: 'PROPERTY_TREE_ADD_PROPERTIES',
+  removePropertyOwners: 'PROPERTY_TREE_REMOVE_OWNERS',
+  removeProperties: 'PROPERTY_TREE_REMOVE_PROPERTIES',
+  refreshGroups: 'GROUPS_REFRESH',
+
+  updateTime: "TIME_UPDATE",
+  subscribeToTime: 'TIME_SUBSCRIBE',
+  unsubscribeToTime: 'TIME_UNSUBSCRIBE',
 
   // action types for connection
   startConnection: 'CONNECTION_START',
@@ -17,10 +22,12 @@ export const actionTypes = {
   onCloseConnection: 'CONNECTION_ON_CLOSE',
   changeConnectionWait: 'CONNECTION_CHANGE_WAIT',
 
-  // action types for fetching data from json files
+  // action types for fetching story data from json files
   fetchData: 'FETCH_DATA',
   fetchDataDone: 'FETCH_DATA_DONE',
   fetchDataFailed: 'FETCH_DATA_FAILED',
+
+  initializeLuaApi: 'LUA_API_INITIALIZE',
 
   // action types for version
   getVersion: 'VERSION_GET',
@@ -30,6 +37,26 @@ export const actionTypes = {
   subscribeToShortcuts: 'SHORTCUTS_SUBSCRIBE',
   unsubscribeToShortcuts: 'SHORTCUTS_UNSUBSCRIBE',
   initializeShortcuts: 'SHORTCUTS_INITIALIZE',
+  executeShortcut: 'SHORTCUT_EXECUTE',
+
+  // action types for session recording
+  subscribeToSessionRecording: 'SESSION_RECORDING_SUBSCRIBE',
+  unsubscribeToSessionRecording: 'SESSION_RECORDING_UNSUBSCRIBE',
+  refreshSessionRecording: 'SESSION_RECORDING_REFRESH',
+  updateSessionRecording: 'SESSION_RECORDING_UPDATE',
+
+  // action types for local gui settings
+  setNavigationAction: 'LOCAL_SET_NAVIGATION_ACTION',
+  setOriginPickerShowFavorites: 'LOCAL_SET_ORIGIN_PICKER_SHOW_FAVORITES',
+  setPropertyTreeExpansion: 'LOCAL_SET_PROPERTY_TREE_EXPANSION',
+  setPopoverPosition: 'LOCAL_POPOVER_SET_POSIITON',
+  setPopoverVisibility: 'LOCAL_POPOVER_SET_VISIBILITY',
+  setPopoverAttachment: 'LOCAL_POPOVER_SET_ATTACHMENT',
+  addNodePropertyPopover: 'LOCAL_ADD_NODE_PROPERTY_POPOVER',
+  removeNodePropertyPopover: 'LOCAL_REMOVE_NODE_PROPERTY_POPOVER',
+  setPopoverActiveTab: 'LOCAL_SET_POPOVER_ACTIVE_TAB',
+
+  setShowAbout: 'LOCAL_SET_SHOW_ABOUT',
 
   // action types for local gui settings
   setNavigationAction: 'LOCAL_SET_NAVIGATION_ACTION',
