@@ -1,14 +1,18 @@
 import { applyMiddleware } from 'redux';
-import { updateBackend } from './propertyTree';
+import { propertyTree } from './propertyTree';
+import { time } from './time';
 import { connection } from './connection';
 import { fetchData } from './fetchData';
 import { shortcuts } from './shortcuts';
 import { version } from './version';
-// import logger from './logger';
+import { sessionRecording } from './sessionRecording';
+import logger from './logger';
 
 const middleware = applyMiddleware(
-  // logger, //middleWare for logging state change
-  updateBackend,
+  //logger, // middleWare for logging state change
+  propertyTree,
+  time,
+  sessionRecording,
   connection,
   shortcuts,
   version,
