@@ -1,68 +1,232 @@
-import DataManager from '../DataManager';
-import { updatePropertyValue, initializePropertyTree } from '../Actions';
-import * as helperFunctions from '../../utils/propertyTreeHelpers';
+import {
+  updatePropertyValue,
+  addPropertyOwners, 
+  addProperties,
+  addProperty,
+  removeProperty,
+  refreshGroups
+} from '../Actions';
 import { actionTypes } from '../Actions/actionTypes';
 import { rootOwnerKey } from '../keys';
 
-const subscriptionIds = [];
+import api from '../api';
 
-const handleUpdatedValues = store => ({ Description, Value }) => {
-  store.dispatch(updatePropertyValue(Description, Value));
-  const state = store.getState();
-  const property = helperFunctions.traverseTreeWithURI(state.propertyTree, Description.Identifier);
-  if (property.listeners < 1) {
-    if (DataManager.unsubscribe(Description.Identifier, subscriptionIds[Description.Identifier])) {
-      delete subscriptionIds[Description.Identifier];
-    }
+import { throttle } from 'lodash/function';
+
+// The property tree middleware is designed to populate the react store's
+// copy of the property tree when the frontend is connected to OpenSpace.
+// The middleware also supports subscribing and setting properties
+// regardless of whether they are present in the redux store or not.
+// However, for values to propagate to the subscribing client,
+// the property needs to exist in the tree.
+// React components retrieve property values from the property tree by using
+// react-redux `connect`.
+//
+// When subscribing to and setting a property, there are four cases:
+// 1) We're connected to the backend. Property exists in the property tree:
+//      This is the happy path.
+//      Subscribing: The subscription is registered and
+//      the subscriber will get updates through redux.
+//      Setting: The data is sent to the backend.
+//
+// 2) We're not connected to backend. Property exists in the property tree:
+//      This means that OpenSpace was once connected but lost connection.
+//      Subscribing: Subscriptions are then stored as `pending`, so they can be
+//      resumed once the connection is established again.
+//      Setting: Properties are considered read-only in disconnected state.
+//      The property will not be set.
+//
+// 3) We're connected to backend. Property does not exist in the property tree:
+//      Subscribing: We are trying to subscribe a property that is not yet loaded
+//      into the frontend, or the property does not exist at all.
+//      There is no way for the frontend to know which of the two cases are true,
+//      so the subscription is stored as `orphan`, and will kick in as soon as the
+//      property can be located in the redux state.
+//      Setting: We will send the update to the backend, which will accept the
+//      setting if the property exists.
+//
+// 4) We're not connected to backend and the property does not exist in the property tree:
+//      Subscribing: We store the subscription as `pending` and will try to promote it to
+//      an active subscription when we are connected.
+//      If it still does not exist, it will be marked `orphan` as in case 3.
+//      Setting: We are not connected, so the property tree is considered read-only.
+//      The property will not be set.
+
+// At this point, we do not support addition and removal
+// of properties in the backend at runtime.
+
+const PendingState = 0;
+const OrphanState = 1;
+const ActiveState = 2;
+
+// Map from uri to { state, nSubscribers, subscription }
+const subscriptionInfos = {};
+
+const handleUpdatedValues = (store, uri, value) => {
+  // Update the value in the redux property tree, based on the
+  // value from the backend.
+  store.dispatch(updatePropertyValue(uri, value));
+
+  // "Lazy unsubscribe":
+  // Cancel the subscription whenever there is an update from the
+  // server, and there are no more active subscibers on the client.
+  // (As opposed to cancelling the subscription immediately when the
+  //  number of subscribers hits zero)
+  const subscriptionInfo = subscriptionInfos[uri];
+  if (subscriptionInfo &&
+      subscriptionInfo.state === ActiveState &&
+      subscriptionInfo.nSubscribers < 1)
+  {
+    subscriptionInfo.subscription.cancel();
+    delete subscriptionInfos[uri];
   }
 };
 
-const startSubscription = (URI, store) => {
-  subscriptionIds[URI] = DataManager.subscribe(URI, handleUpdatedValues(store));
+const tryPromoteSubscription = (store, uri) => {
+  const state = store.getState();
+  const isConnected = state.connection.isConnected;
+  const subscriptionInfo = subscriptionInfos[uri];
+
+  if (!isConnected) {
+    return;
+  }
+
+  if (subscriptionInfo.state === PendingState) {
+    subscriptionInfo.state = OrphanState;
+  }
+
+  const propertyInTree = !!state.propertyTree.properties[uri];
+
+  if (subscriptionInfo.state === OrphanState && propertyInTree) {
+    subscriptionInfo.subscription = createSubscription(store, uri);
+    subscriptionInfo.state = ActiveState;
+  }
 };
 
-const getPropertyTree = (dispatch) => {
-  DataManager.getValue(rootOwnerKey, (Value) => {
-    dispatch(initializePropertyTree(Value));
+const promoteSubscriptions = store => {
+  Object.keys(subscriptionInfos).forEach(uri => {
+    tryPromoteSubscription(store, uri);
   });
 };
 
-const sendDataToBackEnd = (node) => {
-  switch (node.Description.Type) {
-    case 'TransferFunctionProperty': {
-      const convertedEnvelopes = helperFunctions.convertEnvelopes(node.Value);
-      DataManager.setValue(node.Description.Identifier, convertedEnvelopes);
-      break;
+const markAllSubscriptionsAsPending = () => {
+  Object.keys(subscriptionInfos).forEach((uri) => {
+    subscriptionInfos[uri].state = PendingState;
+  });
+};
+
+const createSubscription = (store, uri) => {
+  const subscription = api.subscribeToProperty(uri);
+  const handleUpdates = (value) => handleUpdatedValues(store, uri, value);
+  const throttledHandleUpdates = throttle(handleUpdates, 200);
+
+  (async () => {
+    for await (const data of subscription.iterator()) {
+      throttledHandleUpdates(data.Value);
     }
-    case 'TriggerProperty': {
-      DataManager.trigger(node.Description.Identifier);
-      break;
-    }
-    default: {
-      DataManager.setValue(node.Description.Identifier, node.Value);
-      break;
-    }
+  })();
+  return subscription;
+};
+
+const flattenPropertyTree = (propertyOwner, baseUri) => {
+  let propertyOwners = [];
+  let properties = [];
+  const groups = {};
+
+  propertyOwner.subowners.forEach(subowner => {
+    const uri = baseUri ?
+      baseUri + '.' + subowner.identifier : 
+      subowner.identifier;
+
+    propertyOwners.push({
+      uri,
+      identifier: subowner.identifier,
+      name: subowner.guiName,
+      properties: subowner.properties.map(p => p.Description.Identifier),
+      subowners: subowner.subowners.map(p => uri + '.' + p.identifier),
+      tags: subowner.tag,
+      description: subowner.description
+    });
+    const childData = flattenPropertyTree(subowner, uri);
+    propertyOwners = propertyOwners.concat(childData.propertyOwners);
+    properties = properties.concat(childData.properties);
+  });
+
+  propertyOwner.properties.forEach(property => {
+    const uri = property.Description.Identifier;
+    properties.push({
+      uri,
+      description: property.Description,
+      value: property.Value
+    });
+  });
+
+  return {
+    propertyOwners,
+    properties,
+    groups,
   }
 };
 
-export const updateBackend = store => next => (action) => {
-  const result = next(action);
-  const state = store.getState();
+const getPropertyTree = async (dispatch) => {
+  const value = await api.getProperty(rootOwnerKey);
+  
+  const {propertyOwners, properties, groups} = flattenPropertyTree(value);
+  dispatch(addPropertyOwners(propertyOwners));
+  dispatch(addProperties(properties));
+  dispatch(refreshGroups())
+};
+
+const setBackendValue = (uri, value) => {
+  api.setProperty(uri, value);
+};
+
+export const propertyTree = store => next => (action) => {
+  let result;
+  if (action.type === actionTypes.setPropertyValue) {
+    result = store;
+  } else {
+    result = next(action);
+  }
   switch (action.type) {
-    case actionTypes.onOpenConnection:
+    case actionTypes.onOpenConnection: {
       getPropertyTree(store.dispatch);
       break;
-    case actionTypes.changePropertyTreeNode: {
-      const nodeToUpdate = helperFunctions
-        .traverseTreeWithURI(state.propertyTree, action.payload.URI);
-      sendDataToBackEnd(nodeToUpdate);
+    }
+    case actionTypes.onCloseConnection: {
+      markAllSubscriptionsAsPending();
       break;
     }
-    case actionTypes.startListeningToNode: {
-      const nodeToListen = helperFunctions
-        .traverseTreeWithURI(state.propertyTree, action.payload.URI);
-      if (nodeToListen.listeners === 1) {
-        startSubscription(action.payload.URI, store);
+    case actionTypes.addProperties: {
+      // The added properteis may include properties whose
+      // uri is marked as a `pending`/`orphan` subscription, so
+      // we check if any subscriptions can be promoted to `active`.
+      promoteSubscriptions(store);
+      break;
+    }
+    case actionTypes.setPropertyValue: {
+      setBackendValue(action.payload.uri, action.payload.value);
+      break;
+    }
+    case actionTypes.subscribeToProperty: {
+      const uri = action.payload.uri;
+      const subscriptionInfo = subscriptionInfos[uri];
+      if (subscriptionInfo) {
+        ++subscriptionInfo.nSubscribers;
+      } else {
+        subscriptionInfos[uri] = {
+          state: PendingState,
+          nSubscribers: 1
+        };
+      }
+      tryPromoteSubscription(store, uri);
+      break;
+    }
+    case actionTypes.unsubscribeToProperty: {
+      const uri = action.payload.uri;
+      const subscriptionInfo = subscriptionInfos[uri];
+      if (subscriptionInfo) {
+        --subscriptionInfo.nSubscribers;
       }
       break;
     }
@@ -72,4 +236,4 @@ export const updateBackend = store => next => (action) => {
   return result;
 };
 
-export default updateBackend;
+export default propertyTree;
