@@ -1,17 +1,22 @@
-import { initializeVersion } from '../Actions';
-import { actionTypes } from '../Actions/actionTypes';
-import DataManager from '../DataManager';
+import {
+  initializeVersion,
+} from '../Actions';
 
-const getVersion = callback => {
-  DataManager.getVersion((data) => {
-    callback(data);
-  })
+import api from '../api';
+
+import { actionTypes } from '../Actions/actionTypes';
+
+const getVersion = async callback => {
+  const versionTopic = api.startTopic('version', {});
+  const { value } = await versionTopic.iterator().next();
+  callback(value);
+  versionTopic.cancel();
 };
 
 export const version = store => next => (action) => {
   const result = next(action);
   switch (action.type) {
-    case actionTypes.getVersion:
+    case actionTypes.onOpenConnection:
       getVersion((data) => {
         store.dispatch(initializeVersion(data));
       });
