@@ -1,23 +1,22 @@
-import Connection from '../Connection';
-import DataManager from '../DataManager';
 import { actionTypes } from '../Actions/actionTypes';
 
-export const connection = (state = {}, action) => { // state refers to connection
+const defaultState = {
+  isConnected: false,
+  connectedLost: false,
+}
+
+export const connection = (state = defaultState, action) => { // state refers to connection
   switch (action.type) {
     case actionTypes.startConnection:
       return {
-        connection: DataManager.connection,
-        url: Connection.DEFAULT_URL,
         isConnected: false,
         connectionLost: false,
-        connectionWait: 1000,
       };
     case actionTypes.onOpenConnection:
       return {
         ...state,
         isConnected: true,
         connectionLost: false,
-        connectionWait: 1000,
       };
     case actionTypes.onCloseConnection:
       return {
@@ -31,8 +30,6 @@ export const connection = (state = {}, action) => { // state refers to connectio
         connectionWait: action.payload.value,
       };
     default:
-      return {
-        ...state,
-      };
+      return state;
   }
 };
