@@ -3,13 +3,16 @@ import React from 'react';
 import OriginPicker from './Origin/OriginPicker';
 import TimePicker from './TimePicker';
 import TfEditor from './TfEditor/containers/TfEditor';
+import SessionRec from './SessionRec';
 import styles from './BottomBar.scss';
+
+// <TfEditor /> is currently disabled.
 
 const BottomBar = () => (
   <div className={styles.BottomBar}>
-    <TfEditor />
     <OriginPicker />
     <TimePicker />
+    <SessionRec />
   </div>
 );
 
