@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Icon from '../../common/Icon/Icon';
+import Icon from '../../common/MaterialIcon/MaterialIcon';
 import styles from './Arrows.scss';
 
 const LeftArrow = ({ prevSlide }) => (
