@@ -1,36 +1,38 @@
 import React, { Component } from 'react';
 import PropertyOwner from './Properties/PropertyOwner';
 import Shortcut from './Shortcut';
+import Group from './Group';
+import ContextSection from './ContextSection';
 
 class ScenePaneListItem extends Component {
-    constructor(props) {
-      super(props);
-    }
+  constructor(props) {
+    super(props);
+  }
 
-    render() {
-      var item = this.props;
-      const getMarkup = (propertyOwnerOrShortcut) => {
-        if (item.subowners != null) {
-          return <PropertyOwner
-                  {...propertyOwnerOrShortcut}
-                  key={propertyOwnerOrShortcut.identifier}
-                  onClick={this.props.onSelect}
-                  active={this.props.active}
-               />
-        } else {
-          return <Shortcut
-                  {...propertyOwnerOrShortcut}
-                  key={propertyOwnerOrShortcut.identifier}
-                  onClick={this.props.onSelect}
-                  active={this.props.active}
-                />
-        }
-      };
+  render() {
+    const props = this.props;
 
-      return (
-        getMarkup(item)
-      );
+    if (props.type === 'group') {
+      return <Group key={this.props.path}
+                    path={this.props.path}
+                    expansionIdentifier={'scene/' + this.props.path} />;
     }
+    if (props.type === 'context') {
+      return <ContextSection expansionIdentifier="context" />;
+    }
+    if (props.type === 'propertyOwner') {
+      return <PropertyOwner
+              key={this.props.uri}
+              uri={this.props.uri}
+              expansionIdentifier={'scene-search/' + this.props.uri} />
+    }
+    if (props.type === 'shortcut') {
+      return <Shortcut index={this.props.index}/>
+    }
+    return null;
+  }
+
+
 }
 
 export default ScenePaneListItem;
