@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import SmallLabel from '../../../common/SmallLabel/SmallLabel';
-import Icon from '../../../common/Icon/Icon';
+import Icon from '../../../common/MaterialIcon/MaterialIcon';
 import Popover from '../../../common/Popover/Popover';
 import Picker from '../../../BottomBar/Picker';
 import Button from '../../../common/Input/Button/Button';
@@ -60,7 +61,7 @@ class DateController extends Component {
 
   pickDate(e) {
     this.togglePopover();
-    timeHelpers.setDate(new Date(e.target.id));
+    timeHelpers.setDate(this.props.luaApi, new Date(e.target.id).toISOString());
     const selectedDate = this.props.dateList.find(date => date.date === e.target.id);
     this.props.onChangeSight(selectedDate);
   }
@@ -101,5 +102,13 @@ DateController.defaultProps = {
   onChangeSight: () => {},
   dateList: [],
 };
+
+const mapStateToProps = (state) => {
+  return {
+    luaApi: state.luaApi
+  };
+};
+
+DateController = connect(mapStateToProps)(DateController);
 
 export default DateController;
