@@ -1,7 +1,10 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import styles from './../style/UtilitiesButtons.scss';
 import SmallLabel from '../../../common/SmallLabel/SmallLabel';
+
+import propertyDispatcher from '../../../../api/propertyDispatcher';
 
 class ToggleBoolButton extends Component {
   constructor(props) {
@@ -10,52 +13,60 @@ class ToggleBoolButton extends Component {
     this.toggleProperty = this.toggleProperty.bind(this);
   }
 
-  get propertiesButtons() {
-    return this.props.properties.map((property, i) => {
-      const active = this.props.nodes[i].Value;
-      return (
-        <div
-          className={`${styles.UtilitiesButton} ${active === 'true' && styles.active}`}
-          role="button"
-          tabIndex="0"
-          key={property.URI}
-          onClick={this.toggleProperty}
-          id={property.URI}
-        >
-          <SmallLabel id={property.URI} style={{ textAlign: 'center' }}>
-            {property.label}
-          </SmallLabel>
-        </div>
-      );
-    },
-    );
+  componentDidMount(){
+    this.props.boolPropertyDispatcher.subscribe();
   }
 
-  toggleProperty(e) {
-    this.props.onToggle(e.target.id);
+  componentWillUnMount(){
+    this.props.boolPropertyDispatcher.unsubscribe();
+  }
+
+  toggleProperty() {
+    const value = this.props.propertyNode.value ? false : true;
+    this.props.boolPropertyDispatcher.set(value);
   }
 
   render() {
-    return (
-      <div style={{ display: 'flex' }}>
-        {this.propertiesButtons}
-      </div>
-    );
+      return (
+        <div
+          className={`${styles.UtilitiesButton} ${this.props.propertyNode.value === true && styles.active}`}
+          role="button"
+          tabIndex="0"
+          key={this.props.property.URI}
+          onClick={this.toggleProperty}
+          id={this.props.property.URI}
+        >
+          <SmallLabel id={this.props.property.URI} style={{ textAlign: 'center' }}>
+            {this.props.property.label}
+          </SmallLabel>
+        </div>
+      );
   }
 }
 
-ToggleBoolButton.propTypes = {
-  properties: PropTypes.arrayOf(PropTypes.shape({})),
-  nodes: PropTypes.arrayOf(PropTypes.shape({
-    Value: PropTypes.bool,
-  })),
-  onToggle: PropTypes.func,
-};
+const mapStateToProps = (state, ownProps) => {
 
-ToggleBoolButton.defaultProps = {
-  properties: [],
-  nodes: [],
-  onToggle: () => {},
+  let propertyNode = state.propertyTree.properties[ownProps.property.URI];
+
+  return{
+    propertyNode
+  };
+}
+
+const mapDispatchToProps = (dispatch, ownProps) => ({
+  boolPropertyDispatcher : propertyDispatcher(dispatch, ownProps.property.URI),
+});
+
+ToggleBoolButton = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(ToggleBoolButton);
+
+ToggleBoolButton.propTypes = {
+  propertyNode: PropTypes.shape({
+    description: PropTypes.string,
+    value: PropTypes.bool,
+  }),
 };
 
 export default ToggleBoolButton;
