@@ -10,7 +10,8 @@ class InfoBox extends Component {
     this.state = { showPopup: false };
 
     this.setRef = this.setRef.bind(this);
-    this.togglePopup = this.togglePopup.bind(this);
+    this.showPopup = this.showPopup.bind(this);
+    this.hidePopup = this.hidePopup.bind(this);
   }
 
   setRef(what) {
@@ -25,8 +26,12 @@ class InfoBox extends Component {
     return { top: `${top}px`, left: `${right}px` };
   }
 
-  togglePopup() {
-    this.setState({ showPopup: !this.state.showPopup });
+  showPopup() {
+    this.setState({ showPopup: true });
+  }
+
+  hidePopup() {
+    this.setState({ showPopup: false });
   }
 
   render() {
@@ -35,10 +40,10 @@ class InfoBox extends Component {
     return (
       <span
         ref={this.setRef('wrapper')}
-        onMouseEnter={this.togglePopup}
-        onMouseLeave={this.togglePopup}
       >
-        <MaterialIcon icon={icon} />
+        <MaterialIcon icon={icon}
+         onMouseEnter={this.showPopup}
+         onMouseLeave={this.hidePopup} />
         { showPopup && (
           <Tooltip fixed placement="right" style={this.position}>
             { text }
