@@ -6,17 +6,34 @@ import styles from './Button.scss';
 /* eslint-disable react/no-unused-prop-types */
 
 const Button = (props) => {
-  const specialClasses = 'block small transparent uppercase smalltext nopadding';
+  const specialClasses = 'onClick block small transparent uppercase smalltext nopadding largetext';
   const inheritProps = excludeKeys(props, specialClasses);
 
-  // let extraClass = '';
   const extraClass = specialClasses.split(' ')
     .filter(c => props[c])
     .map(c => styles[c])
     .join(' ');
 
+  let buttonElement = null;
+
+  const onClick = evt => {
+    props.onClick(evt);
+    if (buttonElement) {
+      buttonElement.blur();
+    }
+  };
+
+  const ref = domElement => {
+    buttonElement = domElement;
+  };
+
   return (
-    <button {...inheritProps} className={`${props.className} ${extraClass} ${styles.button}`}>
+    <button
+      {...inheritProps}
+      className={`${props.className} ${extraClass} ${styles.button}`}
+      onClick={onClick}
+      ref={ref}
+    >
       { props.children }
     </button>
   );
@@ -27,6 +44,7 @@ Button.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node.isRequired,
   nopadding: PropTypes.bool,
+  onClick: PropTypes.func,
   small: PropTypes.bool,
   smalltext: PropTypes.bool,
   transparent: PropTypes.bool,
@@ -37,6 +55,7 @@ Button.defaultProps = {
   block: false,
   className: '',
   nopadding: false,
+  onClick: () => {},
   small: false,
   smalltext: false,
   transparent: false,
