@@ -19,6 +19,7 @@ class Input extends Component {
 
     this.onChange = this.onChange.bind(this);
     this.onKeyUp = this.onKeyUp.bind(this);
+    this.onKeyDown = this.onKeyDown.bind(this);
     this.clear = this.clear.bind(this);
     this.setInputRef = this.setInputRef.bind(this);
   }
@@ -44,6 +45,22 @@ class Input extends Component {
   onKeyUp(event) {
     if (event.key === 'Enter') {
       this.props.onEnter(event);
+    }
+  }
+
+  onKeyDown(event) {
+    //TMP hack for windows CEF not allowing .
+    //other issues are not fixed by this but at least it lets .for now.
+    // i.e || (event.keyCode == 39) this should be ' but winds up being right arrow
+    //this solution is so bad it only lets you put . at the end of a value.
+    //but I feel its still better then nothing happening.
+    if ( (window.navigator.platform == "Win32") && ((event.keyCode == 46)) ) {
+      var charString = String.fromCharCode(event.keyCode);
+      if (event.currentTarget.type == "number") {
+        charString += "0";
+      }
+      this.inputNode.value += charString;
+      this.onChange({ currentTarget: this.inputNode });
     }
   }
 
@@ -76,12 +93,12 @@ class Input extends Component {
    * @returns {*}
    */
   get inheritProps() {
-    const doNotInclude = 'onEnter wide onChange loading value clearable';
+    const doNotInclude = 'children onEnter wide onChange loading value clearable';
     return excludeKeys(this.props, doNotInclude);
   }
 
   render() {
-    const { placeholder, className, wide, loading, clearable, label } = this.props;
+    const { placeholder, className, wide, loading, clearable, label, children } = this.props;
     const { value, id } = this.state;
     return (
       <div className={`${styles.group} ${wide ? styles.wide : ''}`}>
@@ -94,12 +111,14 @@ class Input extends Component {
           id={id}
           onChange={this.onChange}
           onKeyUp={this.onKeyUp}
+          onKeyDown={this.onKeyDown}
           value={value}
           ref={this.setInputRef}
         />
         <label htmlFor={id} className={`${styles.label} ${this.hasInput && styles.hasinput}`}>
           { label || placeholder }
         </label>
+        { children }
         { clearable && (
           <MaterialIcon
             icon="cancel"
