@@ -1,26 +1,25 @@
 import {SetDeltaTimeScript, TogglePauseScript, ValuePlaceholder} from "../api/keys";
-import DataManager from "../api/DataManager";
 
-export const togglePause = () => {
-  DataManager.runScript(TogglePauseScript);
+export const togglePause = (luaApi) => {
+  luaApi.time.togglePause();
 };
 
-export const realtime = () => {
-  const script = SetDeltaTimeScript.replace(ValuePlaceholder, 1);
-  DataManager.runScript(script);
+export const realtime = (luaApi) => {
+  luaApi.time.setDeltaTime(1);
 };
 
-export const now = () => {
-  setDate(new Date());
-  UpdateDeltaTimeNow(1);
+export const setDateToNow = (luaApi) => {
+  const now = new Date().toISOString();
+  setDate(luaApi, now);
+  UpdateDeltaTimeNow(luaApi, 1);
 };
 
 // Spice, that is handling the time parsing in OpenSpace does not support
 // ISO 8601-style time zones (the Z). It does, however, always assume that UTC
 // is given.
-export const setDate = (time) => {
-  const fixedTimeString = time.toJSON().replace('Z', '');
-  DataManager.setValue('__time', fixedTimeString);
+export const setDate = (luaApi, time) => {
+  const fixedTimeString = time.toString().replace('Z', '');
+  luaApi.time.setTime(fixedTimeString);
 };
 
 /**
@@ -33,9 +32,8 @@ export const DateStringWithTimeZone = (date, zone = 'Z') => {
   return (!date.includes('Z') ? `${date}${zone}` : date);
 };
 
-export const UpdateDeltaTimeNow = (value) => {
-  const script = SetDeltaTimeScript.replace(ValuePlaceholder, value);
-  DataManager.runScript(script);
+export const UpdateDeltaTimeNow = (luaApi, value) => {
+  luaApi.time.setDeltaTime(value);
 };
 
 export const sortDates = (dateList) => {
